Fix module exports and tests
