+++ melody_generator/__init__.py
@@ -123,21 +123,15 @@ __version__ = "0.1.0"
 #   disable timing jitter for deterministic output.
 # ---------------------------------------------------------------
 
-import mido
-from mido import Message, MidiFile, MidiTrack
 import random
-import sys
 import logging
-import argparse
-from importlib import import_module
 import json
 from pathlib import Path
 from typing import List, Tuple, Optional, Sequence
+from mido import Message, MidiFile, MidiTrack  # noqa: F401
 from functools import lru_cache
 import os
-import math
-import threading
-import re
+import mido  # noqa: F401
 try:
     import numpy as np
 except Exception:  # pragma: no cover - optional dependency
@@ -148,15 +142,15 @@ from .sequence_model import (
     SequenceModel,
     MelodyLSTM,  # noqa: F401
     predict_next,  # noqa: F401
-    load_sequence_model,
+    load_sequence_model,  # noqa: F401
 )  # noqa: F401
-from .style_embeddings import get_style_vector, set_style, get_active_style  # noqa: F401
 # Re-export the full module so ``melody_generator.style_embeddings`` exists for
 # callers that rely on the submodule attribute.  Tests importing the web GUI
 # expect this reference to be present when reloading the package.
 from . import style_embeddings  # noqa: F401
+from .style_embeddings import get_style_vector, set_style, get_active_style  # noqa: F401
 from .tension import tension_for_notes, apply_tension_weights
-from .dynamics import humanize_events
+from .dynamics import humanize_events  # noqa: F401
 from .rhythm_engine import RhythmGenerator, generate_rhythm
 from .feedback import compute_fmd, refine_with_fmd  # noqa: F401
 from .augmentation import (
@@ -678,88 +672,6 @@ def generate_random_rhythm_pattern(length: int = 3) -> List[float]:
     return pattern
 
 
-@lru_cache(maxsize=None)
-def note_to_midi(note: str) -> int:
-    """Convert ``note`` strings such as ``C#4`` into MIDI numbers.
-
-    The octave may contain multiple digits (``C10``) or be negative. A
-    ``ValueError`` is raised for malformed notes.
-
-    Example: ``C#4`` -> ``61`` when ``C4`` equals MIDI note ``60``.
-
-    @param note (str): Note name including octave.
-    @returns int: MIDI note number ``0-127``.
-    Note values outside the standard range are clamped to keep the
-    resulting MIDI pitch valid.
-    """
-    # Extract the note name and octave using a strict pattern.  The pattern
-    # ensures the note consists of a letter with an optional accidental
-    # followed by one or more digits (or a leading minus sign for negative
-    # octaves).
-    # Repeated conversions are automatically memoised by ``lru_cache`` so the
-    # wrapper returns immediately when ``note`` was processed before.
-
-    match = re.fullmatch(r"([A-Ga-g][#b]?)(-?\d+)", note)
-    if not match:
-        logging.error(f"Invalid note format: {note}")
-        raise ValueError(f"Invalid note format: {note}")
-
-    # Normalize case to match the mappings in ``NOTE_TO_SEMITONE`` and convert
-    # the octave to MIDI's numbering scheme (which starts at -1).
-    note_name, octave_str = match.groups()
-    octave = int(octave_str) + 1
-    note_name = note_name.capitalize()
-
-    # Strip the octave number to get the pitch class
-    # and convert flats to their enharmonic sharp equivalents.
-
-    # Map flats to their enharmonic sharps before looking up the semitone index
-    flat_to_sharp = {
-        "Db": "C#",
-        "Eb": "D#",
-        "Fb": "E",
-        "Gb": "F#",
-        "Ab": "G#",
-        "Bb": "A#",
-        "Cb": "B",
-    }
-    note_name = flat_to_sharp.get(note_name, note_name)
-
-    try:
-        # Look up the semitone index within the octave
-        note_idx = NOTE_TO_SEMITONE[note_name]
-    except KeyError:
-        logging.error(f"Note {note_name} not recognized.")
-        raise
-
-    # MIDI note number is calculated relative to C0 at index 12. The resulting
-    # value may fall outside the ``0-127`` range when the octave is extreme,
-    # so clamp it to keep the result valid for standard MIDI files.
-    midi_val = note_idx + (octave * 12)
-    return max(0, min(127, midi_val))
-
-
-def midi_to_note(midi_note: int) -> str:
-    """Convert a MIDI note number back to a note string.
-
-    @param midi_note (int): MIDI note number ``0-127``.
-    @returns str: Note string using sharps (e.g. ``C#4``).
-    """
-    octave = midi_note // 12 - 1
-    # Wrap around the NOTES list to get the pitch class
-    name = NOTES[midi_note % 12]
-    return f"{name}{octave}"
-
-
-def get_interval(note1: str, note2: str) -> int:
-    """Get the interval between two notes in semitones.
-
-    @param note1 (str): First note (e.g., ``C4``).
-    @param note2 (str): Second note (e.g., ``E4``).
-    @returns int: Interval in semitones.
-    """
-    # Interval is the absolute difference in semitone numbers
-    return abs(note_to_midi(note1) - note_to_midi(note2))
 
 
 def get_chord_notes(chord: str) -> List[str]:
@@ -1474,558 +1386,20 @@ def generate_counterpoint_melody(melody: List[str], key: str) -> List[str]:
         prev_base = base_note
     return counter
 
+from .note_utils import note_to_midi, midi_to_note, get_interval  # noqa: E402
+from . import midi_io  # noqa: F401,E402
+from .midi_io import create_midi_file, _open_default_player  # noqa: F401,E402
 
-def create_midi_file(
-    melody: List[str],
-    bpm: int,
-    time_signature: Tuple[int, int],
-    output_file: str,
-    harmony: bool = False,
-    pattern: Optional[List[float]] = None,
-    extra_tracks: Optional[List[List[str]]] = None,
-    chord_progression: Optional[List[str]] = None,
-    chords_separate: bool = True,
-    program: int = 0,
-    humanize: bool = True,
-) -> None:
-    """Create a MIDI file for the given melody and optional chord progression.
+def run_cli():
+    from .cli import run_cli as _run_cli
+    _run_cli()
 
-    ``chord_progression`` may be supplied to render the underlying harmony
-    either on a separate track (default) or on the melody track when
-    ``chords_separate`` is ``False``.
+def main():
+    from .cli import main as _main
+    _main()
 
-    @param melody (List[str]): Sequence of note names representing the melody.
-    @param bpm (int): Beats per minute.
-    @param time_signature (Tuple[int, int]): ``(numerator, denominator)`` pair.
-        The denominator must be one of ``1, 2, 4, 8`` or ``16`` so durations
-        align with standard note values.
-    @param output_file (str): Destination file path.
-    @param harmony (bool): Include a simple harmony line.
-    @param pattern (List[float]|None): Optional rhythmic pattern. If provided it
-        must contain at least one duration and all values must be non-negative;
-        otherwise ``ValueError`` is raised to avoid invalid MIDI timing during
-        event generation.
-    @param extra_tracks (List[List[str]]|None): Additional melody lines.
-    @param chord_progression (List[str]|None): Chords rendered as blocks.
-    @param chords_separate (bool): Write chords to a new track when ``True``.
-    @param program (int): General MIDI instrument program for the melody.
-    @param humanize (bool): Randomly adjust note timing and velocity when ``True``.
-    @returns None: MIDI data is written to ``output_file``.
-    """
-    # ``chord_progression`` must contain at least one chord when provided.
-    # An empty list would lead to ``IndexError`` when aligning chord events
-    # with the melody, so validate early.
-    if chord_progression is not None and not chord_progression:
-        raise ValueError("chord_progression must contain at least one chord")
-    valid_denoms = {1, 2, 4, 8, 16}
-    if time_signature[0] <= 0 or time_signature[1] <= 0 or time_signature[1] not in valid_denoms:
-        raise ValueError(
-            "time_signature denominator must be one of 1, 2, 4, 8 or 16 and numerator must be > 0"
-        )
-    ticks_per_beat = 480
-    mid = MidiFile(ticks_per_beat=ticks_per_beat)
-    track = MidiTrack()
-    mid.tracks.append(track)
-    harmony_track = None
-    if harmony:
-        harmony_track = MidiTrack()
-        mid.tracks.append(harmony_track)
-    extra_midi_tracks: List[MidiTrack] = []
-    if extra_tracks:
-        # Pre-create one MIDI track per extra melody line so that each
-        # additional voice remains isolated in the final file.
-        for _ in extra_tracks:
-            t = MidiTrack()
-            mid.tracks.append(t)
-            extra_midi_tracks.append(t)
-
-    chord_track: Optional[MidiTrack] = None
-    if chord_progression:
-        chord_track = track if not chords_separate else MidiTrack()
-        if chords_separate:
-            mid.tracks.append(chord_track)
-
-    # Set tempo and time signature.
-    track.append(mido.MetaMessage("set_tempo", tempo=mido.bpm2tempo(bpm)))
-    track.append(
-        mido.MetaMessage(
-            "time_signature", numerator=time_signature[0], denominator=time_signature[1]
-        )
-    )
-    track.append(Message("program_change", program=program, time=0))
-
-    # Select a rhythmic pattern and compute the duration of a whole note in ticks.
-    if pattern is None:
-        # Use the rhythm generator to create an onset pattern matching the
-        # melody length when no explicit pattern is provided.
-        pattern = generate_rhythm(len(melody))
-    elif not pattern:
-        # Using an empty pattern would make ``pattern[i % len(pattern)]`` fail
-        # later in the function. Validate early and report a clear error.
-        raise ValueError("pattern must not be empty")
-    elif any(p < 0 for p in pattern):
-        # Negative durations would create invalid delta times in the MIDI file,
-        # so reject them before processing.
-        raise ValueError("pattern durations must be non-negative")
-    whole_note_ticks = ticks_per_beat * 4
-    beat_fraction = 1 / time_signature[1]
-    beat_ticks = int(beat_fraction * whole_note_ticks)
-    beats_per_segment = time_signature[0]
-    beats_elapsed = 0
-    start_beat = 0.0
-    rest_ticks = 0
-    last_note = None
-    last_velocity = 64
-
-    total_beats = 0.0
-
-    # Generate MIDI events for each note using the rhythmic pattern cyclically.
-    for i, note in enumerate(melody):
-        # Determine how long this note should last based on the pattern
-        duration_fraction = pattern[i % len(pattern)]
-        # Create a gentle crescendo followed by a decrescendo so the
-        # phrase has more expressive dynamics than a simple linear ramp.
-        half = max(len(melody) - 1, 1) / 2
-        if i <= half:
-            frac = i / half
-        else:
-            frac = 1 - ((i - half) / half)
-        strong = abs(start_beat - round(start_beat)) < 1e-6
-        velocity = int(50 + 40 * frac)
-        if strong:
-            velocity = min(velocity + 10, 127)
-
-        if duration_fraction == 0:
-            # Treat ``0`` as a rest equal to one beat.
-            rest_ticks += beat_ticks
-            beats_elapsed += 1
-            start_beat += 1
-            continue
-
-        note_duration = int(duration_fraction * whole_note_ticks)
-        midi_note = note_to_midi(note)
-
-        note_on = Message("note_on", note=midi_note, velocity=velocity, time=rest_ticks)
-        note_off = Message("note_off", note=midi_note, velocity=velocity, time=note_duration)
-        track.append(note_on)
-        track.append(note_off)
-        if harmony_track is not None:
-            harmony_note = min(midi_note + 4, 127)
-            h_on = Message(
-                "note_on",
-                note=harmony_note,
-                velocity=max(velocity - 10, 40),
-                time=rest_ticks,
-            )
-            h_off = Message(
-                "note_off",
-                note=harmony_note,
-                velocity=max(velocity - 10, 40),
-                time=note_duration,
-            )
-            harmony_track.append(h_on)
-            harmony_track.append(h_off)
-        # Write notes for any additional melody lines in lockstep with the main
-        # melody. ``zip`` pairs each line with its dedicated MIDI track.
-        for line, t in zip(extra_tracks or [], extra_midi_tracks):
-            if i >= len(line):
-                continue
-            m = note_to_midi(line[i])
-            x_on = Message("note_on", note=m, velocity=velocity, time=rest_ticks)
-            x_off = Message("note_off", note=m, velocity=velocity, time=note_duration)
-            t.append(x_on)
-            t.append(x_off)
-
-        rest_ticks = 0
-        beat_len = duration_fraction / beat_fraction
-        beats_elapsed += beat_len
-        total_beats += beat_len
-        last_note = midi_note
-        last_velocity = velocity
-        start_beat += beat_len
-
-        if beats_elapsed >= beats_per_segment:
-            if last_note is not None and random.random() < 0.5:
-                extra_fraction = random.choice([0.5, 1.0])
-                extra_ticks = int(extra_fraction * whole_note_ticks)
-                on = Message("note_on", note=last_note, velocity=last_velocity, time=0)
-                off = Message("note_off", note=last_note, velocity=last_velocity, time=extra_ticks)
-                track.append(on)
-                track.append(off)
-                rest_ticks = beat_ticks
-            beats_elapsed = 0
-
-    if chord_track is not None:
-        # Compute when each chord should start and end using absolute ticks so
-        # merged chords align with the melody when ``chords_separate`` is False.
-        ticks_per_measure = int(time_signature[0] * ticks_per_beat * (4 / time_signature[1]))
-        ticks_per_chord = ticks_per_measure
-        total_ticks = int(total_beats * ticks_per_beat * (4 / time_signature[1]))
-        num_chords = max(1, math.ceil(total_ticks / ticks_per_chord))
-
-        chord_events: List[Tuple[int, Message]] = []
-        for i in range(num_chords):
-            start_tick = i * ticks_per_chord
-            chord = chord_progression[i % len(chord_progression)]
-            notes = CHORDS.get(chord, [])
-            for note in notes:
-                note_num = note_to_midi(note + "3")
-                chord_events.append(
-                    (start_tick, Message("note_on", note=note_num, velocity=60, time=0))
-                )
-                chord_events.append(
-                    (
-                        start_tick + ticks_per_chord,
-                        Message("note_off", note=note_num, velocity=60, time=0),
-                    )
-                )
-
-        if chords_separate:
-            # Convert absolute times to delta times relative to the chord track.
-            chord_events.sort(key=lambda p: p[0])
-            last = 0
-            for tick, msg in chord_events:
-                msg.time = tick - last
-                chord_track.append(msg)
-                last = tick
-        else:
-            # Merge chord events with the melody track by sorting all events on
-            # their absolute time. First convert existing messages to absolute
-            # ticks so chords can be inserted at the start of the piece.
-            merged_events: List[Tuple[int, Message]] = []
-            current = 0
-            for msg in track:
-                current += msg.time
-                merged_events.append((current, msg))
-            merged_events.extend(chord_events)
-            merged_events.sort(key=lambda p: p[0])
-
-            # Rewrite the melody track with corrected delta times.
-            track.clear()
-            last = 0
-            for tick, msg in merged_events:
-                msg.time = tick - last
-                track.append(msg)
-                last = tick
-
-    # Apply humanization only when requested so callers can produce fully
-    # quantized output for deterministic unit tests or strict sequencing.
-    if humanize:
-        # Operate on the primary melody track; chord tracks remain untouched so
-        # their timing stays predictable when tests assert exact values.
-        humanize_events(mid.tracks[0])
-
-    # Write all tracks to disk
-    mid.save(output_file)
-    logging.info(f"MIDI file saved to {output_file}")
-
-
-def _open_default_player(path: str, *, delete_after: bool = False) -> None:
-    """Launch ``path`` asynchronously with the system's default MIDI player."""
-
-    from .playback import open_default_player
-
-    def runner() -> None:
-        try:
-            open_default_player(path, delete_after=delete_after)
-        except Exception as exc:  # pragma: no cover - platform dependent
-            logging.error("Could not open MIDI file: %s", exc)
-
-    threading.Thread(target=runner, daemon=True).start()
-
-
-def run_cli() -> None:
-    """Parse command line arguments and generate a melody.
-
-    This is a thin wrapper around the core library functions that mirrors
-    the options available in the GUI. It validates the provided arguments,
-    invokes :func:`generate_melody` and finally writes the resulting MIDI
-    file using :func:`create_midi_file`.
-
-    The ``--soundfont`` argument allows supplying a custom ``.sf2`` file
-    used when previewing the result via ``--play``. ``--enable-ml`` loads
-    a small LSTM so note weighting uses learned statistics. ``--style``
-    selects a predefined style embedding.  Pass ``--no-humanize`` to
-    disable timing adjustments so output remains fully quantized for
-    deterministic tests or sequencer import.
-
-    @returns None: Exits via ``sys.exit`` on failure.
-    """
-    # Allow users to quickly inspect supported keys or chords without providing
-    # the rest of the required arguments. These flags are checked before the
-    # ``ArgumentParser`` is instantiated so that missing required options do not
-    # trigger an error when the sole intention is to list available values.
-    if "--list-keys" in sys.argv[1:]:
-        print("\n".join(sorted(SCALE.keys())))
-        return
-    if "--list-chords" in sys.argv[1:]:
-        print("\n".join(sorted(CHORDS.keys())))
-        return
-    parser = argparse.ArgumentParser(
-        description="Generate a random melody and save as a MIDI file."
-    )
-    parser.add_argument(
-        "--list-keys",
-        action="store_true",
-        help="List all supported keys and exit",
-    )
-    parser.add_argument(
-        "--list-chords",
-        action="store_true",
-        help="List all supported chords and exit",
-    )
-    parser.add_argument("--key", type=str, required=True, help="Musical key (e.g., C, Dm, etc.).")
-    parser.add_argument(
-        "--chords", type=str, help="Comma-separated chord progression (e.g., C,Am,F,G)."
-    )
-    parser.add_argument(
-        "--random-chords",
-        type=int,
-        metavar="N",
-        help="Generate a random chord progression of N chords, ignoring --chords.",
-    )
-    parser.add_argument("--bpm", type=int, required=True, help="Beats per minute (integer).")
-    parser.add_argument(
-        "--timesig",
-        type=str,
-        required=True,
-        help="Time signature in numerator/denominator format (e.g., 4/4).",
-    )
-    parser.add_argument("--notes", type=int, required=True, help="Number of notes in the melody.")
-    parser.add_argument("--output", type=str, required=True, help="Output MIDI file path.")
-    parser.add_argument(
-        "--motif_length",
-        type=int,
-        default=4,
-        help="Length of the initial motif (default: 4).",
-    )
-    parser.add_argument("--harmony", action="store_true", help="Add a simple harmony track.")
-    parser.add_argument(
-        "--random-rhythm",
-        action="store_true",
-        help="Generate a random rhythmic pattern.",
-    )
-    parser.add_argument("--counterpoint", action="store_true", help="Generate a counterpoint line.")
-    parser.add_argument(
-        "--harmony-lines",
-        type=int,
-        default=0,
-        metavar="N",
-        help="Number of harmony lines to add in parallel",
-    )
-    parser.add_argument(
-        "--base-octave",
-        type=int,
-        default=4,
-        help=(f"Starting octave for the melody ({MIN_OCTAVE}-{MAX_OCTAVE}, default: 4)."),
-    )
-    parser.add_argument(
-        "--include-chords",
-        action="store_true",
-        help="Add the chord progression to the MIDI output",
-    )
-    parser.add_argument(
-        "--chords-same-track",
-        action="store_true",
-        help="Write chords on the melody track instead of a new one",
-    )
-    parser.add_argument(
-        "--instrument",
-        type=int,
-        default=0,
-        help="MIDI program number for the melody instrument",
-    )
-    parser.add_argument(
-        "--no-humanize",
-        dest="humanize",
-        action="store_false",
-        help="Disable timing and velocity randomization",
-    )
-    parser.add_argument(
-        "--enable-ml",
-        action="store_true",
-        help="Activate ML-based weighting using a small sequence model",
-    )
-    parser.add_argument(
-        "--style",
-        type=str,
-        help="Optional style name to bias note selection",
-    )
-    parser.add_argument(
-        "--soundfont",
-        type=str,
-        help="Path to a SoundFont (.sf2) file used when previewing with --play",
-    )
-    parser.add_argument(
-        "--play", action="store_true", help="Play the MIDI file after it is created"
-    )
-    # Parse the provided CLI arguments
-    args = parser.parse_args()
-
-    # Basic numeric sanity checks
-    if args.bpm <= 0:
-        logging.error("BPM must be a positive integer.")
-        sys.exit(1)
-    if args.notes <= 0:
-        logging.error("Number of notes must be a positive integer.")
-        sys.exit(1)
-    if args.motif_length <= 0:
-        logging.error("Motif length must be a positive integer.")
-        sys.exit(1)
-    if args.motif_length > args.notes:
-        logging.error("Motif length cannot exceed the number of notes.")
-        sys.exit(1)
-
-    # Verify the instrument program falls within the General MIDI range.
-    if not 0 <= args.instrument <= 127:
-        logging.error("Instrument must be between 0 and 127.")
-        sys.exit(1)
-
-    # Ensure the base octave stays within a musically reasonable range so
-    # generated notes remain valid MIDI values.
-    if not MIN_OCTAVE <= args.base_octave <= MAX_OCTAVE:
-        logging.error(f"Base octave must be between {MIN_OCTAVE} and {MAX_OCTAVE}.")
-        sys.exit(1)
-
-    # Validate key and chord progression in a case-insensitive manner so
-    # users can supply values like "c" or "am" without matching the exact
-    # capitalization used by :data:`SCALE` and :data:`CHORDS`.
-    try:
-        args.key = canonical_key(args.key)
-    except ValueError:
-        logging.error("Invalid key provided.")
-        sys.exit(1)
-
-    if args.random_chords:
-        # Let the helper pick a progression based solely on the key
-        chord_progression = generate_random_chord_progression(args.key, args.random_chords)
-    else:
-        if not args.chords:
-            logging.error("Chord progression required unless --random-chords is used.")
-            sys.exit(1)
-        chord_names = [chord.strip() for chord in args.chords.split(",")]
-        chord_progression = []
-        for chord in chord_names:
-            try:
-                chord_progression.append(canonical_chord(chord))
-            except ValueError:
-                logging.error(f"Invalid chord in progression: {chord}")
-                sys.exit(1)
-
-    try:
-        # Parse "numerator/denominator" into two integers and enforce
-        # a denominator drawn from common note values.
-        parts = args.timesig.split("/")
-        if len(parts) != 2:
-            raise ValueError
-        numerator, denominator = map(int, parts)
-        if numerator <= 0 or denominator not in {1, 2, 4, 8, 16}:
-            raise ValueError
-    except ValueError:
-        logging.error(
-            "Time signature must be in the form 'numerator/denominator' with "
-            "numerator > 0 and denominator one of 1, 2, 4, 8 or 16."
-        )
-        sys.exit(1)
-
-    if args.style:
-        try:
-            get_style_vector(args.style)
-        except KeyError:
-            logging.error(f"Unknown style: {args.style}")
-            sys.exit(1)
-
-    seq_model = None
-    if args.enable_ml:
-        try:
-            seq_model = load_sequence_model(None, len(SCALE[args.key]))
-        except RuntimeError as exc:
-            logging.error(str(exc))
-            sys.exit(1)
-
-    rhythm = (
-        generate_random_rhythm_pattern(args.notes) if args.random_rhythm else None
-    )
-    melody = generate_melody(
-        args.key,
-        args.notes,
-        chord_progression,
-        motif_length=args.motif_length,
-        base_octave=args.base_octave,
-        sequence_model=seq_model,
-        style=args.style,
-        pattern=rhythm,
-    )
-    # Collect additional harmony tracks requested on the command line
-    extra: List[List[str]] = []
-    for _ in range(max(0, args.harmony_lines)):
-        # Generate any requested parallel harmony lines
-        extra.append(generate_harmony_line(melody))
-    if args.counterpoint:
-        extra.append(generate_counterpoint_melody(melody, args.key))
-    create_midi_file(
-        melody,
-        args.bpm,
-        (numerator, denominator),
-        args.output,
-        harmony=args.harmony,
-        pattern=rhythm,
-        extra_tracks=extra,
-        chord_progression=chord_progression if args.include_chords else None,
-        chords_separate=not args.chords_same_track,
-        program=args.instrument,
-        humanize=args.humanize,
-    )
-    if args.play:
-        try:
-            from . import playback
-
-            playback.play_midi(args.output, soundfont=args.soundfont)
-        except Exception:
-            _open_default_player(args.output)
-    logging.info("Melody generation complete.")
-
-
-def main() -> None:
-    """Application entry point.
-
-    Runs the CLI when command line arguments are present or launches the
-    Tkinter GUI otherwise.
-
-    @returns None: This function does not return.
-    """
-    # Configure logging only when executing the application directly
-    logging.basicConfig(level=logging.INFO, format="%(levelname)s: %(message)s")
-
-    # Decide between CLI and GUI based on whether arguments were supplied
-    if len(sys.argv) > 1:
-        run_cli()
-    else:
-        try:
-            # Import the GUI lazily so tests without Tkinter can still run
-            MelodyGeneratorGUI = import_module("melody_generator.gui").MelodyGeneratorGUI
-        except ImportError:
-            logging.error(
-                "Tkinter is required for the GUI. Run with CLI arguments or install Tkinter."
-            )
-            logging.error("Tkinter is not available. Please run with CLI options or install it.")
-            sys.exit(1)
-
-        gui = MelodyGeneratorGUI(
-            generate_melody,
-            create_midi_file,
-            SCALE,
-            CHORDS,
-            load_settings,
-            save_settings,
-            generate_random_chord_progression,
-            generate_random_rhythm_pattern,
-            generate_harmony_line,
-            generate_counterpoint_melody,
-        )
-        # Start the Tk event loop
-        gui.run()
 
 
 if __name__ == "__main__":
     main()
+
